Backend/requirements.txt
# API

# Settings & env

# DB (async)
SQLAlchemy[asyncio]>=2.0
asyncpg>=0.29

# Auth & security
passlib[bcrypt]>=1.7
bcrypt>=4.0
PyJWT>=2.9

# Migrations
alembic>=1.13

# Workers / cache 
celery>=5.4
redis>=5.0

# HTTP client for weather/geocoding etc.
httpx[http2]>=0.27

# Utilities
tenacity>=9.0
typing_extensions>=4.12

fastapi==0.119.1
starlette==0.48.0
pydantic==2.9.2
pydantic_core==2.23.4
uvicorn==0.38.0
asyncpg==0.29.0
bcrypt==4.1.2
python-dotenv
aiofiles==25.1.0
requests==2.32.3